Remove tasty pie
