+++ perma_web/api/tests/test_current_user_resource.py
@@ -23,9 +23,6 @@ class CurrentUserResourceTestCase(ApiResourceTransactionTestCase):
             'top_level_folders',
         ]
 
-    def test_get_schema_json(self):
-        self.successful_get(self.detail_url + 'schema/', user=self.org_user)
-
     def test_get_self_detail_json(self):
         self.successful_get(self.detail_url, user=self.org_user, fields=self.fields)
 
@@ -34,7 +31,7 @@ class CurrentUserResourceTestCase(ApiResourceTransactionTestCase):
         self.successful_get(self.detail_url + 'archives/', user=self.regular_user, count=11)
 
     def test_get_folders_json(self):
-        self.successful_get(self.detail_url + 'folders/', user=self.org_user, count=1)
+        self.successful_get(self.detail_url + 'folders/', user=self.org_user, count=2)
 
     def test_get_orgs_json(self):
         self.successful_get(self.detail_url + 'organizations/', user=self.org_user, count=1)
