+++ perma_web/fabfile/dev.py
@@ -52,7 +52,7 @@ def run_ssl(port="0.0.0.0:8000"):
     """
     local("python manage.py runsslserver %s" % port)
 
-_default_tests = "perma api api2 functional_tests lockss"
+_default_tests = "perma api functional_tests lockss"
 
 @task
 def test(apps=_default_tests):
