+++ perma_web/fixtures/api_keys.json
@@ -1,7 +1,7 @@
 [
   {
     "pk": 1,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 1,
       "key": "1d136212a7d2018d53a4e53a16dc7c1e5e1d33f4",
@@ -10,7 +10,7 @@
   },
   {
     "pk": 2,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 2,
       "key": "fdjaskljklfasdjkalfjdklsajkfldjskalfjdks",
@@ -19,7 +19,7 @@
   },
   {
     "pk": 3,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 3,
       "key": "aajklajfdkslajafksdjweeeifjdaizledajk243",
@@ -28,7 +28,7 @@
   },
   {
     "pk": 4,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 4,
       "key": "azjewjvmdaewiofjvewioqr30942fdjeoaewjvjz",
@@ -37,7 +37,7 @@
   },
   {
     "pk": 5,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 5,
       "key": "35f2490vs29riewoafiwri09z32i09sig9490053",
@@ -46,7 +46,7 @@
   },
   {
     "pk": 6,
-    "model": "tastypie.ApiKey",
+    "model": "perma.ApiKey",
     "fields": {
       "user": 6,
       "key": "zjklrfew9f9032fioepwm93i90fjew9090vejerj",
