+++ perma_web/perma/templates/archive/includes/upload_your_own.html
@@ -16,9 +16,6 @@
       <div class="modal-body">
         <form id="archive_upload_form" method="post">
 
-          {# can be removed when Tastypie is no longer used #}
-          <input type="hidden" name="file_required" value="true">
-
           {% if link.guid %}
             <div class="modal-actions">
               <div class="form-group">
