+++ perma_web/perma/templates/archive/single-link.html
@@ -1,5 +1,5 @@
 {% extends "archive/base-archive-responsive.html" %}
-{% load local_datetime as_json archive_description %}
+{% load local_datetime archive_description %}
 {% load render_bundle from webpack_loader %}
 
 {% block header_scripts %}
