+++ perma_web/static/js/create-link.module.js
@@ -82,10 +82,6 @@ function uploadNot (jqXHR) {
   $('.has-error').removeClass('has-error');
 
   if (response) {
-    // Can be removed when Tastypie API no longer used
-    if (response.archives)
-      response = response.archives;
-
     // If error message comes in as {file:"message",url:"message"},
     // show appropriate error message next to each field.
     for (var key in response) {
